requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Analyzer (face landmarks)
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
mediapipe>=0.10.0
numpy>=1.24.0
